Create High Beam Controller and Fuel Analytics
This is a prototype for a system which ensures the correct usage of high beam light in vehicles and detects fuel leakage, fuel theft and refueling to notify the owner. The program is not the complete version of it and just to give an idea of how the project should work. However, the prototype has been tried and tested.